+++ crates/re_analytics/src/web/mod.rs
@@ -1,0 +1,6 @@
+mod config;
+mod pipeline;
+mod sink;
+
+pub use config::{Config, ConfigError};
+pub use pipeline::{Pipeline, PipelineError};
