+++ crates/re_analytics/src/web/pipeline.rs
@@ -6,7 +6,8 @@
 
 use std::time::Duration;
 
-use crate::{Config, Event, PostHogSink};
+use super::sink::PostHogSink;
+use crate::{Config, Event};
 
 // TODO(cmc): abstract away the concept of a `Pipeline` behind an actual trait when comes the time
 // to support more than just PostHog.
@@ -26,14 +27,11 @@ pub enum PipelineError {
 ///
 /// Flushing of the WAL is entirely left up to the OS page cache, hance the -ish.
 #[derive(Debug)]
-pub(crate) struct Pipeline {}
+pub struct Pipeline {}
 
 impl Pipeline {
-    pub(crate) fn new(
-        _config: &Config,
-        _tick: Duration,
-        _sink: PostHogSink,
-    ) -> Result<Option<Self>, PipelineError> {
+    pub(crate) fn new(_config: &Config, _tick: Duration) -> Result<Option<Self>, PipelineError> {
+        let _sink = PostHogSink::default();
         Ok(None)
     }
