+++ crates/re_analytics/src/web/sink.rs
